+++ src/win32/posix_w32.c
@@ -520,17 +520,17 @@ int p_inet_pton(int af, const char* src, void* dst)
 		struct sockaddr_in6 sin6;
 		struct sockaddr_in sin;
 	} sa;
-	size_t srcsize;
+	int srcsize;
 
 	switch(af)
 	{
 		case AF_INET:
 			sa.sin.sin_family = AF_INET;
-			srcsize = sizeof (sa.sin);
+			srcsize = (int)sizeof(sa.sin);
 		break;
 		case AF_INET6:
 			sa.sin6.sin6_family = AF_INET6;
-			srcsize = sizeof (sa.sin6);
+			srcsize = (int)sizeof(sa.sin6);
 		break;
 		default:
 			errno = WSAEPFNOSUPPORT;
