+++ tests-clar/network/fetchlocal.c
@@ -27,7 +27,7 @@ void test_network_fetchlocal__complete(void)
 	cl_git_pass(git_remote_update_tips(origin));
 
 	cl_git_pass(git_reference_list(&refnames, repo, GIT_REF_LISTALL));
-	cl_assert_equal_i(18, refnames.count);
+	cl_assert_equal_i(18, (int)refnames.count);
 	cl_assert(callcount > 0);
 
 	git_strarray_free(&refnames);
@@ -44,7 +44,7 @@ void test_network_fetchlocal__partial(void)
 	const char *url;
 
 	cl_git_pass(git_reference_list(&refnames, repo, GIT_REF_LISTALL));
-	cl_assert_equal_i(1, refnames.count);
+	cl_assert_equal_i(1, (int)refnames.count);
 
 	url = cl_git_fixture_url("testrepo.git");
 	cl_git_pass(git_remote_add(&origin, repo, GIT_REMOTE_ORIGIN, url));
@@ -55,7 +55,7 @@ void test_network_fetchlocal__partial(void)
 	git_strarray_free(&refnames);
 
 	cl_git_pass(git_reference_list(&refnames, repo, GIT_REF_LISTALL));
-	cl_assert_equal_i(19, refnames.count); /* 18 remote + 1 local */
+	cl_assert_equal_i(19, (int)refnames.count); /* 18 remote + 1 local */
 	cl_assert(callcount > 0);
 
 	git_strarray_free(&refnames);
